pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aa-structures"
dynamic = ["version", "description"]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
authors = [{ name = "Erik Kalkoken", email = "kalkoken87@gmail.com" }]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "allianceauth-app-utils>=1.22",
    "allianceauth>=3.0",
    "dhooks-lite>=1.0",
    "django-eveuniverse>=1.5.2",
    "django-multiselectfield",
    "django-navhelper",
    "pytz!=2022.2",
    "redis-simple-mq>=1.0",
    "humanize>=4.7",
]

[project.urls]
Homepage = "https://gitlab.com/ErikKalkoken/aa-structures"
Documentation = "https://aa-structures.readthedocs.io/en/latest/"
Source = "https://gitlab.com/ErikKalkoken/aa-structures"
Changelog = "https://gitlab.com/ErikKalkoken/aa-structures/-/blob/master/CHANGELOG.md"
Tracker = "https://gitlab.com/ErikKalkoken/aa-structures/-/issues"

[tool.flit.module]
name = "structures"

[tool.isort]
profile = "black"
multi_line_output = 3
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DJANGO",
    "ALLIANCEAUTH",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
known_allianceauth = ["allianceauth", "app_utils"]
known_django = ["django", "django_webtest", "esi", "eveuniverse"]

[tool.pylint.'MASTER']
ignore-patterns = ["__init__.py", "auth_hooks.py", "apps.py", "admin.py"]
ignore-paths = ["^.*/tests/.*$", "^.*/migrations/.*$"]

[tool.pylint.'BASIC']
good-names = ["i", "j", "k", "x", "y", "z", "ex", "id"]

[tool.pylint.'FORMAT']
max-line-length = 120

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "too-many-instance-attributes",
    "too-few-public-methods",
    "imported-auth-user",
    "cyclic-import",
    "fixme",
    "import-outside-toplevel",
    "redefined-builtin",
    "no-member",
]
